src/api/requirements.txt
fastapi
fastapi[standard]
requests
gunicorn==21.2.0
azure-cosmos
azure-search-documents==11.4.0
azure-keyvault-secrets
azure-monitor-opentelemetry-exporter
aiohttp==3.9.5
python-dotenv==1.0.1
azure-cognitiveservices-speech==1.37.0
prompty[azure]==0.1.24
jsonlines
nbconvert
pandas
azure-monitor-opentelemetry-exporter
opentelemetry-instrumentation-fastapi
jupyter
opentelemetry-instrumentation
azure-identity==1.17.1
tabulate
langchain-core==0.3.9
langchain-openai==0.2.2
langchain-prompty==0.1.0
